Use up-to-date configuration reference in callbacks
**Summary**
Callbacks are registered during `afterInit`, however they were using a fixed reference to the `dragData` object using the chart instance object at time of initialization. That's fine *only* if the configuration options never change. However if, as was the case for me, the user is changing their callback definitions and updating the options object accordingly, the new callbacks would never be used by this plugin.

**Test Plan**
- Instantiate a chart with one set of callbacks
- Rerender the chart with a new set of callbacks
- Plugin uses the most recent callbacks